+++ src/Integration/Monolog/CloudWatch/src/CloudWatchLogsHandler.php
@@ -98,23 +98,9 @@ class CloudWatchLogsHandler extends AbstractProcessingHandler
     public function __construct(
         CloudWatchLogsClient $client,
         $config,
-        /*$stream,
-        $batchSize = 10000,*/
         $level = Logger::DEBUG,
         $bubble = true
     ) {
-        if (!\is_array($config)) {
-            @\trigger_error('Creating CloudWatchLogsHandler with inline config arguments is deprecated', \E_USER_WARNING);
-            $arguments = \func_get_args();
-            $config = [
-                'group' => $arguments[1],
-                'stream' => $arguments[2],
-                'batchSize' => $arguments[3] ?? null,
-            ];
-            $level = $arguments[4] ?? Logger::DEBUG;
-            $bubble = $arguments[5] ?? true;
-        }
-
         $config['batchSize'] = $config['batchSize'] ?? 10000;
 
         if ($config['batchSize'] > 10000) {